Add DevBootcamper and refactor Student, Teacher, ApprenticeTeacher to use it
