+++ student.rb
@@ -1,16 +1,7 @@
-class Student
-  attr_reader :age, :phase
-  attr_accessor :name
+require_relative 'dev_bootcamper'
 
-  def initialize(options = {})
-    @phase = 1
-    @age = options.fetch(:age, 0)
-    @name = options.fetch(:name, "")
-  end
-
-  def offer_high_five
-    "High five!"
-  end
+class Student < DevBootcamper
+  PHASE_DEFAULT = 1
 
   def set_phase(num)
     response = ""
